following the programing manual more

+++ LSFR.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <stddef.h>
-int32_t LFSR(int32_t reg,int32_t N){
-  int32_t taps=0xB8;
-  for(int i=0;i<N;i++){
+int8_t LFSR(int32_t reg,int32_t N){
+  int8_t taps=0xB8;
+  for(int8_t i=0;i<N;i++){
     printf("reg: %d\n",reg);
-    int32_t res=taps&reg;
-    int32_t feedback_bit=((0b1<<7)&res)^((0b1<<5)&res)^((0b1<<4)&res)^((0b1<<3)&res)^((0b1<<0)&res);
+    int8_t res=taps&reg;
+    int8_t feedback_bit=((0b1<<7)&res)^((0b1<<5)&res)^((0b1<<4)&res)^((0b1<<3)&res)^((0b1<<0)&res);
     reg=reg>>1;
     reg |=(feedback_bit<<7);
   
@@ -15,6 +15,6 @@ int32_t LFSR(int32_t reg,int32_t N){
 }
 int main(){
 
-  int32_t reg=LFSR(9,0xF);
+  int8_t reg=LFSR(9,0xF);
   return 0;
 }
